\encoding{UTF-8}
\name{mefaspat}
\alias{mefaspat}
\docType{package}
\title{
mefaspat: Functions for Spatial Ecological Data and Compatibility with the mefa Package
}
\description{
Miscellaneous functions for mainly multivariate spatial ecological data and compatibility with the mefa package.
}
\details{
To be added later.
}
\author{
Maintainer: \enc{Péter Sólymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\keyword{ package }
\seealso{
\code{\link[mefa:mefa-package]{mefa}}
}
