\name{mefaspat}
\alias{mefaspat}
\docType{package}
\title{
mefaspat - contributions for sedaR project by Peter Solymos
}
\description{
mefaspat - contribution for sedaR package by Peter Solymos}
\details{
\tabular{ll}{
Package: \tab mefaspat\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2008-06-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Maintainer: Peter Solymos <psolymos@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link[mefa:mefa-package]{mefa}}
}
