\name{mefaspat}
\alias{mefaspat}
\docType{package}
\title{
mefaspat - functions for spatial ecological data and compatibility with the mefa package
}
\description{
Miscellaneous functions for mainly multivariate spatial ecological data and compatibility with the mefa package.
}
\details{
Main functionalities: and overview, vignettes ?
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Maintainer: Peter Solymos <psolymos@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link[mefa:mefa-package]{mefa}}
}
